Bot-Telegram-Backup/requirements.txt
python-dotenv
gspread
pandas
telebot
requests
unidecode
spotipy